Refatoração, criação de classe pra formatar string

+++ src/controllers/gameController.ts
@@ -3,21 +3,7 @@ import Game     from "../model/gameModel";
 import GameList from "../model/gameListModel";
 import User from "../model/userModel";
 import FormatDate from "../utils/formatDate";
-
-type Invite = {
-  _id: string;
-  user: {
-    _id: string;
-    confirmed: boolean;
-  };
-}
-
-type User = {
-  _id: string;
-  name: string; 
-  email: string;
-  confirmed: boolean;
-}
+import FormatStrings from "../utils/formatStrings";
 
 export default {
 
@@ -26,18 +12,30 @@ export default {
       const games = await Game.find();
       const gamesInfo = [];
 
-      for(let i = 0; i < games.length; i++){
-        const {_id, name, type, location, description, value, host_ID, date} = games[i];
+      for (let game in games) {
+        const {_id, name, type, location, description, value, host_ID, gameList_ID, date} = games[game];
+        const gameList = await GameList.findOne({_id: gameList_ID});
+
+        const fiveDays = 5 * 24 * 60 * 60 * 1000;
+        let now = Number(Date.now());
+        let gameDate = Number(new Date(date)) + fiveDays;
+
+        if (gameDate < now) {
+          games[game].delete();
+          gameList.delete();
+          continue;
+        }
 
         gamesInfo.push({
-          _id, name, type, location, description, value, host_ID, 
-          date: FormatDate.toDateString(date), hour: FormatDate.hourToString(date)
+          _id, name, type, location, description, value: FormatStrings.formatMoneyToUser(value), host_ID,
+          date: FormatDate.toDateString(date), hour: FormatDate.hourToString(date),
+          gameList
         });
       }
 
       res.status(200).json(gamesInfo);
     } catch(error){
-      res.status(500).json({message: 'Ops! Something went wrong!'})
+      res.status(500).json({message: error.message});
     }
   },
 
@@ -46,23 +44,30 @@ export default {
     try{
       let {name, type, location, description, value, host_ID, date, hour} = req.body;
       let formatDate = FormatDate.dateToDatabase(date, hour);
-      
+
       let game = new Game({
-        name, type, location, description, value, date: formatDate, host_ID, hour
+        name, type, location, description, value: FormatStrings.formatMoneyToDatabase(value), date: formatDate, host_ID, hour
       });
-      
+
       let gameList = new GameList({
         game_ID: game._id, host_ID
       });
 
       game.gameList_ID = gameList._id;
-      
+
+      let now = Number(Date.now());
+      let gameDate = Number(new Date(date));
+
+      if (gameDate < now) {
+        res.status(200).json({error: "The event date cannot be less than the current date"});
+        return;
+      }
+
       await game.save();
       await gameList.save();
 
-      res.status(200).send({game, gameList})//}, gameList});
+      res.status(200).send({game, gameList});
     } catch(error){
-      console.log(error);
       res.status(500).json({message: "Ops! Something went wrong"});
     }
 
@@ -73,33 +78,34 @@ export default {
       let {id} = req.params;
       const game = await Game.findOne({_id: id});
       const gameList = await GameList.findOne({_id: game.gameList_ID});
-      const invitedUsers = new Array<User>();
+      const invitedUsers = new Array();
 
-      for(let i = 0; i < gameList.invitedUsers.length; i++){
-        let user = await User.findOne({_id: gameList.invitedUsers[i].user._id});
-        
-        const invitedUser: User = {
+      for(let invited in gameList.invitedUsers){
+        let user = await User.findOne({_id: gameList.invitedUsers[invited].user._id});
+
+        const invitedUser = {
           _id: user._id,
           name: user.name,
           email: user.email,
-          confirmed: gameList.invitedUsers[i].user.confirmed
+          confirmed: gameList.invitedUsers[invited].user.confirmed
         }
-        
+
         invitedUsers.push(invitedUser);
       }
 
-      const {name, type, location, description, value, host_ID, date} = game;
+      const {_id, name, type, location, description, value, host_ID, gameList_ID, date} = game;
 
       const gameInfo = {
-        name, type, location, description, value, host_ID, 
-        date: FormatDate.toDateString(date), hour: FormatDate.hourToString(date)
+        _id,
+        name, type, location, description, value, host_ID, gameList_ID,
+        date: FormatDate.toDateString(date), hour: FormatDate.hourToString(date),
+        gameList: {
+          invitedUsers: invitedUsers
+        }
       }
 
-      res.status(200).json({gameInfo, gameList:{
-        invitedUsers
-      }});
-    } catch(error){
-      console.log(error);
+      res.status(200).json(gameInfo);
+    } catch(error) {
       res.status(500).json({message: "Ops! Something went wrong"});
     }
   },
@@ -119,7 +125,7 @@ export default {
 
       gameList.invitedUsers.push(user);
       gameList.save();
-      
+
       res.status(200).json({gameList});
     } catch(error){
       res.status(500).json({message: "Ops! Something went wrong"});
@@ -132,14 +138,14 @@ export default {
 
       const gameList = await GameList.findOne({_id: gameListId});
 
-      gameList.invitedUsers.forEach((user: Invite) => {
-        if(user._id == inviteId){
-          user.user.confirmed = true;
+      for (let invited in gameList.invitedUsers) {
+        if(gameList.invitedUsers[invited]._id == inviteId){
+          gameList.invitedUsers[invited].user.confirmed = true;
         }
-      });
-      
+      }
+
       gameList.save();
-      
+
       res.status(200).json({gameList});
     } catch(error){
       res.status(500).json({message: "Ops! Something went wrong"});
@@ -150,17 +156,17 @@ export default {
     try{
       let {userId} = req.params;
 
-      const gameList = await GameList.find();
-      const invitations = new Array<Invite>();
+      const gameLists = await GameList.find();
+      const invitations = new Array();
 
-      for(let i = 0; i < gameList.length; i++){
-        gameList[i].invitedUsers.forEach((user: Invite) => {
-          if(user.user._id == userId){
-            invitations.push(user);
+      for (let gameList in gameLists) {
+        for (let invited in gameLists[gameList].invitedUsers) {
+          if (gameLists[gameList].invitedUsers[invited].user._id == userId) {
+            invitations.push(gameLists[gameList].invitedUsers[invited].user);
           }
-        });
+        }
       }
-      
+
       res.status(200).json({invitations});
     } catch(error){
       res.status(500).json({message: "Ops! Something went wrong"});
@@ -170,18 +176,22 @@ export default {
   async destroy(req: Request, res: Response){
     try{
       const {id} = req.params;
+      const {host_ID} = req.body
 
       const game = await Game.findOne({_id: id});
       const gameList = await GameList.findOne({_id: game.gameList_ID});
+
+      if (game.host_ID !== host_ID) {
+        return res.status(401).json({message: "Only the event creator can delete it"});
+      }
 
       game.delete();
       gameList.delete();
 
       res.status(200).json({message: "Game deleted successfully"});
-    } catch(error){
+    } catch(error) {
       console.log(error);
       res.status(500).json({message: "Ops! Something went wrong"});
     }
   }
-  
 }
